requirements.txt
Flask==2.3.3
supabase==1.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
flask-cors==4.0.0